Map refactor
Each user has a color for its targets in order to make it easy to visualize data as admin. 
The Google API key is a sensible variable so it was placed in the .env file.
As all the points will be in the United States the code which center de map in browser current location was removed and the map is just centered in the United States.